% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoa.R
\name{pcoa}
\alias{pcoa}
\title{PCoA Implementation}
\usage{
pcoa(dist_matrix, k = 3)
}
\arguments{
\item{dist_matrix}{numeric matrix indicating the pairwise distance matrix}

\item{k}{numeric variable indicating the number of dimension coordinate system (e.g. k = 2 or 3 space), default is 3}
}
\value{
the PCoA coordinates matrix
}
\description{
PCoA Implementation
}
\examples{
data(dist_matrix)
pcoa(dist_matrix)
pcoa(dist_matrix, k = 3)
}
